@@ -26,15 +26,13 @@
  THE SOFTWARE.
  */
 
-#include <boost/program_options.hpp>
-namespace po = boost::program_options;
-using std::string;
-using std::vector;
-
+#include <string>
+#include <vector>
 #if defined(__GNUC__) && defined(__linux__)
-#include <fenv.h>
+#include <cfenv>
 #endif
-#include <signal.h>
+#include <cstdint>
+#include <set>
 #include <gmp.h>
 
 #include "time_mem.h"
@@ -43,22 +41,20 @@ using std::vector;
 #include <cryptominisat5/dimacsparser.h>
 #include <cryptominisat5/streambuffer.h>
 #include <arjun/arjun.h>
+#include "src/argparse.hpp"
 
 using namespace CMSat;
 using std::cout;
 using std::cerr;
 using std::endl;
-ApproxMC::AppMC* appmc = NULL;
-ArjunNS::Arjun* arjun = NULL;
+using std::set;
+using std::string;
+using std::vector;
+ApproxMC::AppMC* appmc = nullptr;
+ArjunNS::Arjun* arjun = nullptr;
+argparse::ArgumentParser program = argparse::ArgumentParser("approxmc");
 
-po::options_description main_options = po::options_description("Main options");
-po::options_description arjun_options = po::options_description("Arjun options");
-po::options_description improvement_options = po::options_description("Improvement options");
-po::options_description misc_options = po::options_description("Misc options");
-po::options_description help_options;
-po::variables_map vm;
-po::positional_options_description p;
-uint32_t verbosity = 1;
+uint32_t verb = 1;
 uint32_t seed;
 string randfilename;
 double epsilon;
@@ -73,20 +69,29 @@ uint32_t reuse_models = 1;
 uint32_t force_sol_extension = 0;
 uint32_t sparse = 0;
 int dump_intermediary_cnf = 0;
-uint32_t must_mult_exp2 = 0;
 
 //Arjun
-vector<uint32_t> sampling_vars;
-bool sampling_vars_found = false;
 int ignore_sampl_set = 0;
 int do_arjun = 1;
 int debug_arjun = 0;
 int debug = 0;
 int with_e = 1;
 int e_iter_1 = 2;
-int e_iter_2 = 2;
-int e_vivif_sparsify = 0;
+int e_iter_2 = 0;
+int e_vivif = 1;
+int e_sparsify = 0;
 int e_get_reds = 0;
+
+#define myopt(name, var, fun, hhelp) \
+    program.add_argument(name) \
+        .action([&](const auto& a) {var = std::fun(a.c_str());}) \
+        .default_value(var) \
+        .help(hhelp)
+#define myopt2(name1, name2, var, fun, hhelp) \
+    program.add_argument(name1, name2) \
+        .action([&](const auto& a) {var = std::fun(a.c_str());}) \
+        .default_value(var) \
+        .help(hhelp)
 
 void add_appmc_options()
 {
@@ -98,208 +103,84 @@ void add_appmc_options()
     sparse = tmp.get_sparse();
     seed = tmp.get_seed();
 
-    std::ostringstream my_epsilon;
-    std::ostringstream my_delta;
-    std::ostringstream my_var_elim_ratio;
 
-    my_epsilon << std::setprecision(8) << epsilon;
-    my_delta << std::setprecision(8) << delta;
-    my_var_elim_ratio << std::setprecision(8) << var_elim_ratio;
+    myopt2("-v", "--verb", verb, atoi, "Verbosity");
+    myopt2("-s", "--seed", seed, atoi, "Seed");
+    myopt2("-e", "--epsilon", epsilon, stod,
+            "Tolerance parameter, i.e. how close is the count from the correct count? "
+            "Count output is within bounds of (exact_count/(1+e)) < count < (exact_count*(1+e)). "
+            "So e=0.8 means we'll output at most 180%% of exact count and at least 55%% of exact count. "
+            "Lower value means more precise.");
+    myopt2("-d", "--delta", delta, stod, "Confidence parameter, i.e. how sure are we of the result? "
+            "(1-d) = probability the count is within range as per epsilon parameter. "
+            "So d=0.2 means we are 80%% sure the count is within range as specified by epsilon. "
+            "The lower, the higher confidence we have in the count.");
+    myopt("--ignore", ignore_sampl_set, atoi, "Ignore given sampling set and recompute it with Arjun");
+    myopt("--randbits", randfilename, string, "Read random bits from this file.");
+    myopt("--cert", certfilename, string, "Put certification of ApproxMC execution to this file.");
 
-    main_options.add_options()
-    ("help,h", "Prints help")
-    ("input", po::value< vector<string> >(), "file(s) to read")
-    ("verb,v", po::value(&verbosity)->default_value(1), "verbosity")
-    ("seed,s", po::value(&seed)->default_value(seed), "Seed")
-    ("randbits", po::value(&randfilename), "file to read random bits")
-    ("version", "Print version info")
+    /* arjun_options.add_options() */
+    myopt("--arjun", do_arjun, atoi, "Use arjun to minimize sampling set");
+    myopt("--arjundebug", debug_arjun, atoi, "Use CNF from Arjun, but use sampling set from CNF");
 
-    ("epsilon,e", po::value(&epsilon)->default_value(epsilon, my_epsilon.str())
-        , "Tolerance parameter, i.e. how close is the count from the correct count? Count output is within bounds of (exact_count/(1+e)) < count < (exact_count*(1+e)). So e=0.8 means we'll output at most 180%% of exact count and at least 55%% of exact count. Lower value means more precise.")
-    ("delta,d", po::value(&delta)->default_value(delta, my_delta.str())
-        , "Confidence parameter, i.e. how sure are we of the result? (1-d) = probability the count is within range as per epsilon parameter. So d=0.2 means we are 80%% sure the count is within range as specified by epsilon. The lower, the higher confidence we have in the count.")
-    ("log", po::value(&logfilename),
-         "Logs of ApproxMC execution")
-    ("cert", po::value(&certfilename),
-         "Certification of ApproxMC execution")
-    ("ignore", po::value(&ignore_sampl_set)->default_value(ignore_sampl_set)
-        , "Ignore given sampling set and recompute it with Arjun")
-    ("debug", po::value(&debug)->default_value(debug), "Turn on more heavy internal debugging")
-    ;
+    /* improvement_options.add_options() */
+    myopt("--sparse", sparse, atoi,
+            "0 = (default) Do not use sparse method. 1 = Generate sparse XORs when possible.");
+    myopt("--reusemodels", reuse_models, atoi, "Reuse models while counting solutions");
+    myopt("--forcesolextension", force_sol_extension, atoi,
+            "Use trick of not extending solutions in the SAT solver to full solution");
+    myopt("--withe", with_e, atoi, "Eliminate variables and simplify CNF as well");
+    myopt("--eiter1", e_iter_1, atoi, "Num iters of E on 1st round");
+    myopt("--eiter2", e_iter_2, atoi, "Num iters of E on 1st round");
+    myopt("--evivif", e_vivif, atoi, "E vivif");
+    myopt("--esparsif", e_sparsify, atoi, "E sparsify");
+    myopt("--egetreds", e_get_reds, atoi, "Get redundant from E");
 
-    ArjunNS::Arjun tmpa;
-    arjun_options.add_options()
-    ("arjun", po::value(&do_arjun)->default_value(do_arjun)
-        , "Use arjun to minimize sampling set")
-    ("arjundebug", po::value(&debug_arjun)->default_value(debug_arjun)
-        , "Use CNF from Arjun, but use sampling set from CNF")
-    ;
+    /* misc_options.add_options() */
+    myopt("--verbcls", verb_cls, atoi, "Print banning clause + xor clauses. Highly verbose.");
+    myopt("--simplify", simplify, atoi, "Simplify aggressiveness");
+    myopt("--velimratio", var_elim_ratio, stod, "Variable elimination ratio for each simplify run");
+    myopt("--dumpintercnf", dump_intermediary_cnf, atoi,
+            "Dump intermediary CNFs during solving into files cnf_dump-X.cnf. If set to 1 only UNSAT is dumped, if set to 2, all are dumped");
+    myopt("--log", logfilename, string, "Put logs of ApproxMC execution to this file");
+    myopt("--debug", debug, atoi, "Turn on more heavy internal debugging");
 
-    improvement_options.add_options()
-    ("sparse", po::value(&sparse)->default_value(sparse)
-        , "0 = (default) Do not use sparse method. 1 = Generate sparse XORs when possible.")
-    ("reusemodels", po::value(&reuse_models)->default_value(reuse_models)
-        , "Reuse models while counting solutions")
-    ("forcesolextension", po::value(&force_sol_extension)->default_value(force_sol_extension)
-        , "Use trick of not extending solutions in the SAT solver to full solution")
-    ("withe", po::value(&with_e)->default_value(with_e)
-        , "Eliminate variables and simplify CNF as well")
-    ("eiter1", po::value(&e_iter_1)->default_value(e_iter_1)
-        , "Num iters of E on 1st round")
-    ("eiter2", po::value(&e_iter_2)->default_value(e_iter_2)
-        , "Num iters of E on 1st round")
-    ("evivifsparsify", po::value(&e_vivif_sparsify)->default_value(e_vivif_sparsify)
-        , "E vivif+sparsify")
-    ("egetreds", po::value(&e_get_reds)->default_value(e_get_reds)
-        , "Get redundant from E")
-    ;
-
-    misc_options.add_options()
-    ("verbcls", po::value(&verb_cls)->default_value(verb_cls)
-        ,"Print banning clause + xor clauses. Highly verbose.")
-    ("simplify", po::value(&simplify)->default_value(simplify)
-        , "Simplify aggressiveness")
-    ("velimratio", po::value(&var_elim_ratio)->default_value(var_elim_ratio)
-        , "Variable elimination ratio for each simplify run")
-    ("dumpintercnf", po::value(&dump_intermediary_cnf)->default_value(dump_intermediary_cnf)
-        , "Dump intermediary CNFs during solving into files cnf_dump-X.cnf. If set to 1 only UNSAT is dumped, if set to 2, all are dumped");
-    ;
-
-    help_options.add(main_options);
-    help_options.add(improvement_options);
-    help_options.add(arjun_options);
-    help_options.add(misc_options);
+    program.add_argument("inputfile").remaining().help("input CNF");
 }
 
-void add_supported_options(int argc, char** argv)
-{
+void add_supported_options(int argc, char** argv) {
     add_appmc_options();
-    p.add("input", 1);
-
     try {
-        po::store(po::command_line_parser(argc, argv).options(help_options).positional(p).run(), vm);
-        if (vm.count("help"))
-        {
-            cout
-            << "Probably Approximate counter" << endl;
-
-            cout
-            << "approxmc [options] inputfile" << endl << endl;
-
-            cout << help_options << endl;
-            std::exit(0);
+        program.parse_args(argc, argv);
+        if (program.is_used("--help")) {
+            cout << "Probilistic Approcimate Counter" << endl << endl
+            << "approxmc [options] inputfile" << endl;
+            cout << program << endl;
+            exit(0);
         }
-
-        if (vm.count("version")) {
-            cout << appmc->get_version_info();
-            std::exit(0);
-        }
-
-        po::notify(vm);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::unknown_option> >& c
-    ) {
-        cerr
-        << "ERROR: Some option you gave was wrong. Please give '--help' to get help" << endl
-        << "       Unknown option: " << c.what() << endl;
-        std::exit(-1);
-    } catch (boost::bad_any_cast &e) {
-        std::cerr
-        << "ERROR! You probably gave a wrong argument type" << endl
-        << "       Bad cast: " << e.what()
-        << endl;
-
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::invalid_option_value> >& what
-    ) {
-        cerr
-        << "ERROR: Invalid value '" << what.what() << "'" << endl
-        << "       given to option '" << what.get_option_name() << "'"
-        << endl;
-
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::multiple_occurrences> >& what
-    ) {
-        cerr
-        << "ERROR: " << what.what() << " of option '"
-        << what.get_option_name() << "'"
-        << endl;
-
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::required_option> >& what
-    ) {
-        cerr
-        << "ERROR: You forgot to give a required option '"
-        << what.get_option_name() << "'"
-        << endl;
-
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::too_many_positional_options_error> >& what
-    ) {
-        cerr
-        << "ERROR: You gave too many positional arguments. Only the input CNF can be given as a positional option." << endl;
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::ambiguous_option> >& what
-    ) {
-        cerr
-        << "ERROR: The option you gave was not fully written and matches" << endl
-        << "       more than one option. Please give the full option name." << endl
-        << "       The option you gave: '" << what.get_option_name() << "'" <<endl
-        << "       The alternatives are: ";
-        for(size_t i = 0; i < what.alternatives().size(); i++) {
-            cout << what.alternatives()[i];
-            if (i+1 < what.alternatives().size()) {
-                cout << ", ";
-            }
-        }
-        cout << endl;
-
-        std::exit(-1);
-    } catch (boost::exception_detail::clone_impl<
-        boost::exception_detail::error_info_injector<po::invalid_command_line_syntax> >& what
-    ) {
-        cerr
-        << "ERROR: The option you gave is missing the argument or the" << endl
-        << "       argument is given with space between the equal sign." << endl
-        << "       detailed error message: " << what.what() << endl
-        ;
-        std::exit(-1);
+    }
+    catch (const std::exception& err) {
+        std::cerr << err.what() << std::endl;
+        exit(-1);
     }
 
+    if (program["version"] == true) {
+        cout << appmc->get_version_info();
+        exit(0);
+    }
 }
 
-// void SIGINT_handler(int)
-// {
-//     if (!appmc) {
-//         return;
-//     }
-//     SolCount solCount = appmc->calc_est_count();
-//     if (!solCount.valid) {
-//         cout << "c did not manage to get a single measurement, we have no estimate of the count" << endl;
-//         exit(-1);
-//     }
-//     cout << "c Below count is NOT FULLY APPROXIMIATE due to early-abort!" << endl;
-//     solCount.print_num_solutions();
-//     exit(-1);
-// }
-
-template<class T>
-void read_in_file(const string& filename, T* myreader)
+template<class T> void read_in_file(const string& filename, T* myreader)
 {
     #ifndef USE_ZLIB
     FILE * in = fopen(filename.c_str(), "rb");
-    DimacsParser<StreamBuffer<FILE*, FN>, T> parser(myreader, NULL, verbosity);
+    DimacsParser<StreamBuffer<FILE*, FN>, T> parser(myreader, nullptr, verb);
     #else
     gzFile in = gzopen(filename.c_str(), "rb");
-    DimacsParser<StreamBuffer<gzFile, GZ>, T> parser(myreader, NULL, verbosity);
+    DimacsParser<StreamBuffer<gzFile, GZ>, T> parser(myreader, nullptr, verb);
     #endif
 
-    if (in == NULL) {
+    if (in == nullptr) {
         std::cerr
         << "ERROR! Could not open file '"
         << filename
@@ -310,10 +191,6 @@ void read_in_file(const string& filename, T* myreader)
 
     if (!parser.parse_DIMACS(in, true)) exit(-1);
 
-    sampling_vars = parser.sampling_vars;
-    sampling_vars_found = parser.sampling_vars_found;
-    must_mult_exp2 = parser.must_mult_exp2;
-
     #ifndef USE_ZLIB
     fclose(in);
     #else
@@ -323,10 +200,8 @@ void read_in_file(const string& filename, T* myreader)
 
 inline double stats_line_percent(double num, double total)
 {
-    if (total == 0) {
-        return 0;
-    } else {
-        return num/total*100.0;
+    if (total == 0) { return 0;
+    } else { return num/total*100.0;
     }
 }
 
@@ -350,12 +225,8 @@ void print_final_indep_set(
     << " %" << endl;
 }
 
-template<class T>
-void read_stdin(T* myreader)
-{
-    cout
-    << "c Reading from standard input... Use '-h' or '--help' for help."
-    << endl;
+template<class T> void read_stdin(T* myreader) {
+    cout << "c Reading from standard input... Use '-h' or '--help' for help." << endl;
 
     #ifndef USE_ZLIB
     FILE * in = stdin;
@@ -363,115 +234,78 @@ void read_stdin(T* myreader)
     gzFile in = gzdopen(0, "rb"); //opens stdin, which is 0
     #endif
 
-    if (in == NULL) {
+    if (in == nullptr) {
         std::cerr << "ERROR! Could not open standard input for reading" << endl;
         std::exit(1);
     }
 
     #ifndef USE_ZLIB
-    DimacsParser<StreamBuffer<FILE*, FN>, T> parser(myreader, NULL, verbosity);
+    DimacsParser<StreamBuffer<FILE*, FN>, T> parser(myreader, nullptr, verb);
     #else
-    DimacsParser<StreamBuffer<gzFile, GZ>, T> parser(myreader, NULL, verbosity);
+    DimacsParser<StreamBuffer<gzFile, GZ>, T> parser(myreader, nullptr, verb);
     #endif
 
-    if (!parser.parse_DIMACS(in, false)) {
-        exit(-1);
-    }
-
-    sampling_vars = parser.sampling_vars;
-    sampling_vars_found = parser.sampling_vars_found;
+    if (!parser.parse_DIMACS(in, false)) exit(-1);
 
     #ifdef USE_ZLIB
     gzclose(in);
     #endif
 }
 
-void print_num_solutions(uint32_t cellSolCount, uint32_t hashCount)
+void print_num_solutions(uint32_t cell_sol_cnt, uint32_t hash_count, const mpz_class& mult)
 {
     cout << "c [appmc] Number of solutions is: "
-    << cellSolCount << "*2**" << hashCount << endl;
-    if (cellSolCount == 0) {
-        cout << "s UNSATISFIABLE" << endl;
-    } else {
-        cout << "s SATISFIABLE" << endl;
-    }
+    << cell_sol_cnt << "*2**" << hash_count << "*" << mult << endl;
+    if (cell_sol_cnt == 0) cout << "s UNSATISFIABLE" << endl;
+    else cout << "s SATISFIABLE" << endl;
 
-    mpz_t num_sols;
-    mpz_init (num_sols);
-    mpz_ui_pow_ui(num_sols, 2, hashCount);
-    mpz_mul_ui(num_sols, num_sols, cellSolCount);
+    mpz_class num_sols(2);
+    mpz_pow_ui(num_sols.get_mpz_t(), num_sols.get_mpz_t(), hash_count);
+    num_sols *= cell_sol_cnt;
+    num_sols *= mult;
 
-    cout << "s mc " << std::flush;
-    mpz_out_str(0, 10, num_sols);
-    cout << endl;
-    mpz_clear(num_sols);
-
+    cout << "s mc " << num_sols << endl;
 }
 
 void get_cnf_from_arjun() {
-    bool ret = true;
     const uint32_t orig_num_vars = arjun->get_orig_num_vars();
     appmc->new_vars(orig_num_vars);
-    arjun->start_getting_small_clauses(
-        std::numeric_limits<uint32_t>::max(),
-        std::numeric_limits<uint32_t>::max(),
-        false);
+    arjun->start_getting_constraints();
     vector<Lit> clause;
-    while (ret) {
-        ret = arjun->get_next_small_clause(clause);
-        if (!ret) break;
-
+    bool is_xor, rhs;
+    while (arjun->get_next_constraint(clause, is_xor, rhs)) {
+        assert(!is_xor); assert(rhs);
         bool ok = true;
-        for(auto l: clause) {
-            if (l.var() >= orig_num_vars) { ok = false; break; }
-        }
+        for(auto l: clause) if (l.var() >= orig_num_vars) { ok = false; break; }
         if (ok) appmc->add_clause(clause);
     }
-    arjun->end_getting_small_clauses();
-
-    vector<Lit> lits;
-    for(const auto& bnn: arjun->get_bnns()) {
-        if (bnn) {
-            lits.clear();
-            lits.insert(lits.end(), bnn->begin(), bnn->end());
-            appmc->add_bnn_clause(lits, bnn->cutoff, bnn->out);
-        }
-    }
+    arjun->end_getting_constraints();
 }
 
-template<class T>
-void read_input_cnf(T* reader)
-{
-    //Init Arjun, read in file, get minimal indep set
-    if (vm.count("input") != 0) {
-        vector<string> inp = vm["input"].as<vector<string> >();
-        if (inp.size() > 1) {
-            cout << "[appmc] ERROR: you must only give one CNF as input" << endl;
+template<class T> void read_input_cnf(T* reader) {
+    try {
+        auto files = program.get<std::vector<std::string>>("inputfile");
+        if (files.size() > 1) {
+            cout << "ERROR: you can only pass at most one positional option, an INPUT file" << endl;
             exit(-1);
         }
-
-        read_in_file(inp[0].c_str(), reader);
-    } else {
+        assert(!files.empty());
+        const string inp = files[0];
+        read_in_file(inp, reader);
+    } catch (std::logic_error& e) {
         read_stdin(reader);
     }
-}
-
-uint32_t set_up_sampling_set()
-{
-    uint32_t orig_sampling_set_size;
-    if (!sampling_vars_found || ignore_sampl_set) {
-        orig_sampling_set_size = arjun->start_with_clean_sampling_set();
-    } else {
-        orig_sampling_set_size = arjun->set_starting_sampling_set(sampling_vars);
+    if (!reader->get_sampl_vars_set()  || ignore_sampl_set) {
+        vector<uint32_t> all_vars;
+        for(uint32_t i = 0; i < reader->nVars(); i++) all_vars.push_back(i);
+        reader->set_sampl_vars(all_vars);
     }
-
-    return orig_sampling_set_size;
 }
 
 void set_approxmc_options()
 {
     //Main options
-    appmc->set_verbosity(verbosity);
+    appmc->set_verbosity(verb);
     appmc->set_seed(seed);
     appmc->set_epsilon(epsilon);
     appmc->set_delta(delta);
@@ -493,7 +327,7 @@ void set_approxmc_options()
         appmc->set_dump_intermediary_cnf(std::max(dump_intermediary_cnf, 1));
     }
 
-    if (logfilename != "") {
+    if (!logfilename.empty()) {
         appmc->set_up_log(logfilename);
         cout << "c [appmc] Logfile set " << logfilename << endl;
     }
@@ -540,104 +374,77 @@ void transfer_unit_clauses_from_arjun()
 int main(int argc, char** argv)
 {
     #if defined(__GNUC__) && defined(__linux__)
-    feenableexcept(FE_INVALID   |
-                   FE_DIVBYZERO |
-                   FE_OVERFLOW
-                  );
+    feenableexcept(FE_INVALID   | FE_DIVBYZERO | FE_OVERFLOW);
     #endif
-//     signal(SIGALRM, SIGINT_handler);
-//     signal(SIGTERM, SIGINT_handler);
     double start_time = cpuTime();
 
     //Reconstruct the command line so we can emit it later if needed
     string command_line;
     for(int i = 0; i < argc; i++) {
         command_line += string(argv[i]);
-        if (i+1 < argc) {
-            command_line += " ";
-        }
+        if (i+1 < argc) command_line += " ";
     }
 
     appmc = new ApproxMC::AppMC;
     add_supported_options(argc, argv);
-    if (verbosity) {
+    if (verb) {
         cout << appmc->get_version_info();
         cout << "c executed with command line: " << command_line << endl;
     }
     set_approxmc_options();
 
-    uint32_t offset_count_by_2_pow = 0;
     if (do_arjun) {
         //Arjun-based minimization
         arjun = new ArjunNS::Arjun;
         arjun->set_seed(seed);
-        arjun->set_verbosity(verbosity);
+        arjun->set_verbosity(verb);
         arjun->set_simp(simplify);
-
-        if (verbosity) cout << "c Arjun SHA revision " <<  arjun->get_version_info() << endl;
+        if (verb) cout << "c Arjun SHA revision " <<  arjun->get_version_info() << endl;
 
         read_input_cnf(arjun);
-        print_orig_sampling_vars(sampling_vars, arjun);
-        auto debug_sampling_vars = sampling_vars; // debug ONLY
-        const uint32_t orig_sampling_set_size = set_up_sampling_set();
-        sampling_vars = arjun->get_indep_set();
-        vector<uint32_t> empty_occ_sampl_vars;
-        empty_occ_sampl_vars = arjun->get_empty_occ_sampl_vars();
-        print_final_indep_set(
-            sampling_vars , orig_sampling_set_size, empty_occ_sampl_vars);
+        print_orig_sampling_vars(arjun->get_orig_sampl_vars(), arjun);
+        auto debug_sampling_vars = arjun->get_orig_sampl_vars();
+        auto sampl_vars = arjun->run_backwards();
+        print_final_indep_set(sampl_vars, arjun->get_orig_sampl_vars().size(),
+                arjun->get_empty_sampl_vars());
         if (with_e) {
             ArjunNS::SimpConf sc;
-            sc.oracle_vivify = e_vivif_sparsify;
+            sc.appmc = true;
+            sc.oracle_vivify = e_vivif;
             sc.oracle_vivify_get_learnts = true;
-            sc.oracle_sparsify = e_vivif_sparsify;
+            sc.oracle_sparsify = e_sparsify;
             sc.iter1 = e_iter_1;
             sc.iter2 = e_iter_2;
-            const auto ret = arjun->get_fully_simplified_renumbered_cnf(sampling_vars, sc, true, false);
+            const auto ret = arjun->get_fully_simplified_renumbered_cnf(sc);
             appmc->new_vars(ret.nvars);
             for(const auto& cl: ret.cnf) appmc->add_clause(cl);
             if (e_get_reds) for(const auto& cl: ret.red_cnf) appmc->add_red_clause(cl);
-            sampling_vars = ret.sampling_vars;
-            offset_count_by_2_pow = ret.empty_occs;
+            sampl_vars = ret.sampl_vars;
+            appmc->set_multiplier_weight(ret.multiplier_weight);
         } else {
-            std::set<uint32_t> sampl_vars_set;
-            sampl_vars_set.insert(sampling_vars.begin(), sampling_vars.end());
-            for(auto const& v: empty_occ_sampl_vars) {
-                assert(sampl_vars_set.find(v) != sampl_vars_set.end()); // this is guaranteed by arjun
-                sampl_vars_set.erase(v);
-            }
-            offset_count_by_2_pow = empty_occ_sampl_vars.size();
-            sampling_vars.clear();
-            sampling_vars.insert(sampling_vars.end(), sampl_vars_set.begin(), sampl_vars_set.end());
             get_cnf_from_arjun();
             transfer_unit_clauses_from_arjun();
+            mpz_class dummy(2);
+            mpz_pow_ui(dummy.get_mpz_t(), dummy.get_mpz_t(), arjun->get_empty_sampl_vars().size());
+            appmc->set_multiplier_weight(arjun->get_multiplier_weight()*dummy);
         }
         if (debug_arjun) {
-            sampling_vars = debug_sampling_vars;
-            offset_count_by_2_pow = 0;
+            assert(!with_e && "Can't use debug and --withe at the same time");
+            sampl_vars = debug_sampling_vars;
+            appmc->set_multiplier_weight(1);
         }
+        appmc->set_sampl_vars(sampl_vars);
         delete arjun;
     } else {
         read_input_cnf(appmc);
-        if (ignore_sampl_set || !sampling_vars_found) {
-            sampling_vars.clear();
-            for(uint32_t i = 0; i < appmc->nVars(); i++) sampling_vars.push_back(i);
-        }
-        print_final_indep_set(sampling_vars , 0, vector<uint32_t>());
+        print_final_indep_set(appmc->get_sampl_vars() , 0, vector<uint32_t>());
     }
 
     ApproxMC::SolCount sol_count;
-    if (!sampling_vars.empty()) {
-        appmc->set_projection_set(sampling_vars);
-        sol_count = appmc->count();
-        appmc->print_stats(start_time);
-    } else {
-        bool ret = appmc->find_one_solution();
-        sol_count.hashCount = 0;
-        if (ret) sol_count.cellSolCount = 1;
-        else sol_count.cellSolCount = 0;
-    }
+    sol_count = appmc->count();
+    appmc->print_stats(start_time);
     cout << "c [appmc+arjun] Total time: " << (cpuTime() - start_time) << endl;
-    print_num_solutions(sol_count.cellSolCount, sol_count.hashCount+offset_count_by_2_pow+must_mult_exp2);
+    print_num_solutions(sol_count.cellSolCount, sol_count.hashCount, appmc->get_multiplier_weight());
 
     delete appmc;
 }
