@@ -70,33 +70,33 @@ DLL_PUBLIC AppMC::~AppMC()
 // Helper function, used only in this unit
 void setup_sampling_vars(AppMCPrivateData* data)
 {
-    if (data->conf.sampling_set.empty()) {
+    if (data->conf.sampl_vars.empty()) {
         if (data->conf.verb) {
             cout
             << "c [appmc] WARNING! Sampling set was not declared! We will be **VERY** slow"
             << endl;
         }
         for (size_t i = 0; i < data->counter.solver->nVars(); i++) {
-            data->conf.sampling_set.push_back(i);
+            data->conf.sampl_vars.push_back(i);
         }
     }
 
     if (data->conf.verb) {
-        cout << "c [appmc] Sampling set size: " << data->conf.sampling_set.size() << endl;
-        if (data->conf.sampling_set.size() > 100) {
+        cout << "c [appmc] Sampling set size: " << data->conf.sampl_vars.size() << endl;
+        if (data->conf.sampl_vars.size() > 100) {
             cout
             << "c [appmc] Sampling var set contains over 100 variables, not displaying"
             << endl;
         } else {
             cout << "c [appmc] Sampling set: ";
-            for (auto v: data->conf.sampling_set) {
+            for (auto v: data->conf.sampl_vars) {
                 cout << v+1 << " ";
             }
             cout << "0" << endl;
         }
     }
 
-    data->counter.solver->set_sampling_vars(&(data->conf.sampling_set));
+    data->counter.solver->set_sampl_vars(data->conf.sampl_vars);
 }
 
 DLL_PUBLIC string AppMC::get_version_info()
@@ -244,8 +244,9 @@ DLL_PUBLIC ApproxMC::SolCount AppMC::count()
     return sol_count;
 }
 
-DLL_PUBLIC void AppMC::set_projection_set(const vector<uint32_t>& vars)
+DLL_PUBLIC void AppMC::set_sampl_vars(const vector<uint32_t>& vars)
 {
+    data->conf.sampl_vars_set = true;
     for(const auto& v: vars) {
         if (v >= data->counter.solver->nVars()) {
             std::cout << "ERROR: function set_projection_set() called with variable that is larger than the number of variables inside the solver. Exiting." << endl;
@@ -253,7 +254,11 @@ DLL_PUBLIC void AppMC::set_projection_set(const vector<uint32_t>& vars)
             exit(-1);
         }
     }
-    data->conf.sampling_set = vars;
+    data->conf.sampl_vars = vars;
+}
+
+DLL_PUBLIC const std::vector<uint32_t>& AppMC::get_sampl_vars() const  {
+    return data->conf.sampl_vars;
 }
 
 
@@ -291,18 +296,6 @@ DLL_PUBLIC bool AppMC::add_xor_clause(const vector<uint32_t>& vars, bool rhs)
     return data->counter.solver_add_xor_clause(vars, rhs);
 }
 
-DLL_PUBLIC bool AppMC::add_bnn_clause(
-            const std::vector<CMSat::Lit>& lits,
-            signed cutoff,
-            Lit out)
-{
-    if (data->conf.dump_intermediary_cnf) {
-        cout << "ERROR: BNNs not supported when dumping" << endl;
-        exit(-1);
-    }
-    return data->counter.solver->add_bnn_clause(lits, cutoff, out);
-}
-
 DLL_PUBLIC CMSat::SATSolver* AppMC::get_solver()
 {
     return data->counter.solver;
@@ -310,7 +303,7 @@ DLL_PUBLIC CMSat::SATSolver* AppMC::get_solver()
 
 DLL_PUBLIC const std::vector<uint32_t>& AppMC::get_sampling_set() const
 {
-    return data->conf.sampling_set;
+    return data->conf.sampl_vars;
 }
 
 DLL_PUBLIC void AppMC::set_dump_intermediary_cnf(const int dump_intermediary_cnf)
@@ -327,3 +320,31 @@ DLL_PUBLIC void AppMC::print_stats(const double start_time)
         data->counter.solver->set_verbosity(data->conf.verb);
     }
 }
+
+DLL_PUBLIC bool AppMC::get_sampl_vars_set() const {
+    return data->conf.sampl_vars_set;
+}
+
+ DLL_PUBLIC void AppMC::set_multiplier_weight(const mpz_class& weight) {
+     data->conf.multiplier_weight = weight;
+ }
+
+ DLL_PUBLIC const mpz_class& AppMC::get_multiplier_weight() const {
+     return data->conf.multiplier_weight;
+ }
+
+DLL_PUBLIC void AppMC::set_weighted(const bool weighted) {
+    if (weighted) {
+        cout << "ERROR: Weighted ApproxMC not supported" << endl;
+        exit(-1);
+    }
+}
+
+DLL_PUBLIC void AppMC::set_lit_weight(const Lit&, const double) {
+    cout << "ERROR: Weighted ApproxMC is not supported" << endl;
+    exit(-1);
+}
+
+ DLL_PUBLIC void AppMC::set_opt_sampl_vars(const std::vector<uint32_t>&) {
+     // Not interesting for AppMC
+ }
